adjusted tests location based on mock, fixed random_password method missing
