+++ tests/integration/vsphere/test_Datastore_File.py
