+++ tests/integration/vsphere/test_ESX_Cli.py
