+++ tests/integration/vsphere/test_ESXi_Ssh.py
