+++ tests/integration/vsphere/test_Sdk.py
