+++ tests/integration/vsphere/test_Task.py
