+++ tests/integration/vsphere/test_VM_Device.py
