+++ tests/integration/vsphere/test_VM_Keystroke.py
