+++ tests/integration/vsphere/test_VM_Screenshot.py
