+++ tests/integration/vsphere/test_VM_Task.py
